+++ R/platform.R
@@ -66,14 +66,51 @@ current_r_platform <- function() {
 #' @rdname current_r_platform
 
 current_r_platform_data <- function() {
-  raw <- get_platform()
-  platform <- parse_platform(raw)
-  if (platform$os == "linux" || substr(platform$os, 1, 6) == "linux-") {
-    platform <- current_r_platform_data_linux(platform)
+  forced <- forced_platform()
+  if (!is.null(forced)) {
+    platform <- parse_platform(forced)
+  } else {
+    raw <- get_platform(forced = FALSE)
+    platform <- parse_platform(raw)
+    if (platform$os == "linux" || substr(platform$os, 1, 6) == "linux-") {
+      platform <- current_r_platform_data_linux(platform)
+    }
   }
 
   platform$platform <- apply(platform, 1, paste, collapse = "-")
   platform
+}
+
+valid_platform_string <- function(x) {
+  grepl("^[^-].*[-][^-].*[-][^-].*$", x)
+}
+
+forced_platform <- function() {
+  opt <- getOption("pkg.current_platform")
+  if (!is.null(opt)) {
+    if (!is_string(opt)) {
+      stop("The `pkg.current_platform` option must be a string scalar.")
+    }
+    if (!valid_platform_string(opt)) {
+      stop("The pkg.current_platform` option must be a valid platform ",
+           "triple: `cpu-vendor-os`. \"", opt, "\" is not.")
+    }
+    return(opt)
+  }
+  env <- Sys.getenv("PKG_CURRENT_PLATFORM")
+  if (env != "") {
+    if (is.na(env) || !valid_platform_string(env)) {
+      stop("The `PKG_CURRENT_PLATFORM` environment variable must be a valid ",
+           "platform triple: \"cpu-vendor-os\". \"", env, "\" is not.")
+    }
+    return(env)
+  }
+
+  NULL
+}
+
+get_platform <- function(forced = TRUE) {
+  (if (forced) forced_platform()) %||% R.version$platform
 }
 
 #' @details